handson/serverless/s3/requirements.txt
aws-cdk.aws-s3==1.100.0
boto3
